ensure that custom print scales will be properly set on printpanel instantiation
